// prisma/generated/zod/inputTypeSchemas/BookingRateSnapshotWhereInputSchema.ts
import { Prisma } from '@prisma/client';

import { z } from 'zod';
import { StringFilterSchema } from './StringFilterSchema';
import { StringNullableFilterSchema } from './StringNullableFilterSchema';
import { DecimalFilterSchema } from './DecimalFilterSchema';
import { isValidDecimalInput } from './isValidDecimalInput';
import { DecimalJsLikeSchema } from './DecimalJsLikeSchema';
import { EnumRateDurationTypeFilterSchema } from './EnumRateDurationTypeFilterSchema';
import { RateDurationTypeSchema } from './RateDurationTypeSchema';
import { BoolFilterSchema } from './BoolFilterSchema';
import { DateTimeNullableFilterSchema } from './DateTimeNullableFilterSchema';
import { EnumTaxTypeFilterSchema } from './EnumTaxTypeFilterSchema';
import { TaxTypeSchema } from './TaxTypeSchema';
import { DateTimeFilterSchema } from './DateTimeFilterSchema';
import { RateScalarRelationFilterSchema } from './RateScalarRelationFilterSchema';
import { RateWhereInputSchema } from './RateWhereInputSchema';
import { BookingItemListRelationFilterSchema } from './BookingItemListRelationFilterSchema';

export const BookingRateSnapshotWhereInputSchema: z.ZodType<Prisma.BookingRateSnapshotWhereInput> = z.object({
  AND: z.union([ z.lazy(() => BookingRateSnapshotWhereInputSchema),z.lazy(() => BookingRateSnapshotWhereInputSchema).array() ]).optional(),
  OR: z.lazy(() => BookingRateSnapshotWhereInputSchema).array().optional(),
  NOT: z.union([ z.lazy(() => BookingRateSnapshotWhereInputSchema),z.lazy(() => BookingRateSnapshotWhereInputSchema).array() ]).optional(),
  id: z.union([ z.lazy(() => StringFilterSchema),z.string() ]).optional(),
  rateId: z.union([ z.lazy(() => StringFilterSchema),z.string() ]).optional(),
  name: z.union([ z.lazy(() => StringFilterSchema),z.string() ]).optional(),
  description: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  price: z.union([ z.lazy(() => DecimalFilterSchema),z.union([z.number(),z.string(),z.instanceof(Prisma.Decimal),DecimalJsLikeSchema,]).refine((v) => isValidDecimalInput(v), { message: 'Must be a Decimal' }) ]).optional(),
  currency: z.union([ z.lazy(() => StringFilterSchema),z.string() ]).optional(),
  durationType: z.union([ z.lazy(() => EnumRateDurationTypeFilterSchema),z.lazy(() => RateDurationTypeSchema) ]).optional(),
  isDefault: z.union([ z.lazy(() => BoolFilterSchema),z.boolean() ]).optional(),
  isActive: z.union([ z.lazy(() => BoolFilterSchema),z.boolean() ]).optional(),
  bookingStartDate: z.union([ z.lazy(() => DateTimeNullableFilterSchema),z.coerce.date() ]).optional().nullable(),
  bookingEndDate: z.union([ z.lazy(() => DateTimeNullableFilterSchema),z.coerce.date() ]).optional().nullable(),
  usageStartDate: z.union([ z.lazy(() => DateTimeNullableFilterSchema),z.coerce.date() ]).optional().nullable(),
  usageEndDate: z.union([ z.lazy(() => DateTimeNullableFilterSchema),z.coerce.date() ]).optional().nullable(),
  taxType: z.union([ z.lazy(() => EnumTaxTypeFilterSchema),z.lazy(() => TaxTypeSchema) ]).optional(),
  taxAmount: z.union([ z.lazy(() => DecimalFilterSchema),z.union([z.number(),z.string(),z.instanceof(Prisma.Decimal),DecimalJsLikeSchema,]).refine((v) => isValidDecimalInput(v), { message: 'Must be a Decimal' }) ]).optional(),
  createdAt: z.union([ z.lazy(() => DateTimeFilterSchema),z.coerce.date() ]).optional(),
  updatedAt: z.union([ z.lazy(() => DateTimeFilterSchema),z.coerce.date() ]).optional(),
  rate: z.union([ z.lazy(() => RateScalarRelationFilterSchema),z.lazy(() => RateWhereInputSchema) ]).optional(),
  BookingItem: z.lazy(() => BookingItemListRelationFilterSchema).optional()
}).strict();

export default BookingRateSnapshotWhereInputSchema;